+++ mycodo/outputs/mcp23017.py
@@ -1,7 +1,9 @@
 # coding=utf-8
 #
-# pcf8574.py - Output for PCF8574
+# mcp23017.py - Output for MCP23017
 #
+from collections import OrderedDict
+
 from flask_babel import lazy_gettext
 
 from mycodo.config_translations import TRANSLATIONS
@@ -10,100 +12,34 @@ from mycodo.outputs.base_output import AbstractOutput
 from mycodo.utils.database import db_retrieve_table_daemon
 
 # Measurements
-measurements_dict = {
-    0: {
+measurements_dict = OrderedDict()
+channels_dict = OrderedDict()
+for each_channel in range(16):
+    measurements_dict[each_channel] = {
         'measurement': 'duration_time',
-        'unit': 's',
-    },
-    1: {
-        'measurement': 'duration_time',
-        'unit': 's',
-    },
-    2: {
-        'measurement': 'duration_time',
-        'unit': 's',
-    },
-    3: {
-        'measurement': 'duration_time',
-        'unit': 's',
-    },
-    4: {
-        'measurement': 'duration_time',
-        'unit': 's',
-    },
-    5: {
-        'measurement': 'duration_time',
-        'unit': 's',
-    },
-    6: {
-        'measurement': 'duration_time',
-        'unit': 's',
-    },
-    7: {
-        'measurement': 'duration_time',
-        'unit': 's',
+        'unit': 's'
     }
-}
-
-channels_dict = {
-    0: {
-        'name': 'Channel 1',
+    channels_dict[each_channel] = {
+        'name': 'Channel {}'.format(each_channel + 1),
         'types': ['on_off'],
-        'measurements': [0]
-    },
-    1: {
-        'name': 'Channel 2',
-        'types': ['on_off'],
-        'measurements': [1]
-    },
-    2: {
-        'name': 'Channel 3',
-        'types': ['on_off'],
-        'measurements': [2]
-    },
-    3: {
-        'name': 'Channel 4',
-        'types': ['on_off'],
-        'measurements': [3]
-    },
-    4: {
-        'name': 'Channel 5',
-        'types': ['on_off'],
-        'measurements': [4]
-    },
-    5: {
-        'name': 'Channel 6',
-        'types': ['on_off'],
-        'measurements': [5]
-    },
-    6: {
-        'name': 'Channel 7',
-        'types': ['on_off'],
-        'measurements': [6]
-    },
-    7: {
-        'name': 'Channel 8',
-        'types': ['on_off'],
-        'measurements': [7]
+        'measurements': [each_channel]
     }
-}
 
 # Output information
 OUTPUT_INFORMATION = {
-    'output_name_unique': 'PCF8574',
-    'output_name': "{}: PCF8574 (8 Channels): {}".format(
+    'output_name_unique': 'MCP23017',
+    'output_name': "{}: MCP23017 (16 Channels): {}".format(
         lazy_gettext("I/O Expander"), lazy_gettext('On/Off')),
-    'output_manufacturer': 'Texas Instruments',
-    'output_library': 'smbus2',
+    'output_manufacturer': 'MICROCHIP',
     'measurements_dict': measurements_dict,
     'channels_dict': channels_dict,
     'output_types': ['on_off'],
 
-    'url_manufacturer': 'https://www.ti.com/product/PCF8574',
-    'url_datasheet': 'https://www.ti.com/lit/ds/symlink/pcf8574.pdf',
-    'url_product_purchase': 'https://www.amazon.com/gp/product/B07JGSNWFF',
+    'url_manufacturer': 'https://www.microchip.com/wwwproducts/en/MCP23017',
+    'url_datasheet': 'https://ww1.microchip.com/downloads/en/devicedoc/20001952c.pdf',
+    'url_product_purchase': 'https://www.amazon.com/Waveshare-MCP23017-Expansion-Interface-Expands/dp/B07P2H1NZG',
 
-    'message': 'Controls the 8 channels of the PCF8574.',
+    'message': 'Controls the 16 channels of the MCP23017.',
 
     'options_enabled': [
         'i2c_location',
@@ -113,14 +49,13 @@ OUTPUT_INFORMATION = {
     'options_disabled': ['interface'],
 
     'dependencies_module': [
-        ('pip-pypi', 'smbus2', 'smbus2==0.4.1')
+        ('pip-pypi', 'usb.core', 'pyusb==1.1.1'),
+        ('pip-pypi', 'adafruit_extended_bus', 'adafruit-extended-bus==1.0.1'),
+        ('pip-pypi', 'adafruit_mcp4728', 'adafruit-circuitpython-mcp230xx==2.4.6')
     ],
 
     'interfaces': ['I2C'],
-    'i2c_location': [
-        '0x20', '0x21', '0x22', '0x23', '0x24', '0x25', '0x26', '0x27',
-        '0x38', '0x39', '0x3a', '0x3b', '0x3c', '0x3d', '0x3e', '0x3f'
-    ],
+    'i2c_location': ['0x20', '0x21', '0x22', '0x23', '0x24', '0x25', '0x26', '0x27', ],
     'i2c_address_editable': False,
     'i2c_address_default': '0x20',
 
@@ -191,6 +126,7 @@ class OutputModule(AbstractOutput):
         super(OutputModule, self).__init__(output, testing=testing, name=__name__)
 
         self.sensor = None
+        self.pins = []
 
         output_channels = db_retrieve_table_daemon(
             OutputChannel).filter(OutputChannel.output_id == output.unique_id).all()
@@ -198,40 +134,38 @@ class OutputModule(AbstractOutput):
             OUTPUT_INFORMATION['custom_channel_options'], output_channels)
 
     def setup_output(self):
-        import smbus2
+        from adafruit_mcp230xx.mcp23017 import MCP23017
+        from adafruit_extended_bus import ExtendedI2C
 
         self.setup_output_variables(OUTPUT_INFORMATION)
 
         try:
-            self.logger.debug("I2C: Address: {}, Bus: {}".format(self.output.i2c_location, self.output.i2c_bus))
+            self.logger.debug("I2C: Address: {}, Bus: {}".format(
+                self.output.i2c_location, self.output.i2c_bus))
             if self.output.i2c_location:
-                self.sensor = PCF8574(smbus2, self.output.i2c_bus, int(str(self.output.i2c_location), 16))
+                self.sensor = MCP23017(
+                    ExtendedI2C(self.output.i2c_bus),
+                    address=int(str(self.output.i2c_location), 16))
                 self.output_setup = True
         except:
             self.logger.exception("Could not set up output")
             return
 
-        dict_states = {}
+        for pin in range(0, 16):
+            self.pins.append(self.sensor.get_pin(pin))
+
         for channel in channels_dict:
             if self.options_channels['state_startup'][channel] == 1:
-                dict_states[channel] = bool(self.options_channels['on_state'][channel])
+                self.pins[channel].switch_to_output(value=bool(self.options_channels['on_state'][channel]))
+                self.output_states[channel] = bool(self.options_channels['on_state'][channel])
             elif self.options_channels['state_startup'][channel] == 0:
-                dict_states[channel] = bool(not self.options_channels['on_state'][channel])
+                self.pins[channel].switch_to_output(value=bool(not self.options_channels['on_state'][channel]))
+                self.output_states[channel] = bool(not self.options_channels['on_state'][channel])
             else:
                 # Default state: Off
-                dict_states[channel] = bool(not self.options_channels['on_state'][channel])
+                self.pins[channel].switch_to_output(value=bool(not self.options_channels['on_state'][channel]))
+                self.output_states[channel] = bool(not self.options_channels['on_state'][channel])
 
-        self.logger.debug("List sent to device: {}".format(self.dict_to_list_states(dict_states)))
-        try:
-            self.sensor.port(self.dict_to_list_states(dict_states))
-        except OSError as err:
-            self.logger.error(
-                "OSError: {}. Check that the device is connected properly, the correct "
-                "address is selected, and you can communicate with the device.".format(err))
-
-        self.output_states = dict_states
-
-        for channel in channels_dict:
             if self.options_channels['trigger_functions_startup'][channel]:
                 self.check_triggers(self.unique_id, output_channel=channel)
 
@@ -252,19 +186,15 @@ class OutputModule(AbstractOutput):
             return msg
 
         try:
-            dict_states = {}
-            for channel in channels_dict:
-                if output_channel == channel:
-                    if state == 'on':
-                        dict_states[channel] = bool(self.options_channels['on_state'][channel])
-                    elif state == 'off':
-                        dict_states[channel] = bool(not self.options_channels['on_state'][channel])
-                else:
-                    dict_states[channel] = self.output_states[channel]
+            if state == 'on':
+                self.pins[output_channel].value = bool(self.options_channels['on_state'][output_channel])
+                self.output_states[output_channel] = bool(self.options_channels['on_state'][output_channel])
+                self.logger.debug("Output channel {} turned ON".format(output_channel))
+            elif state == 'off':
+                self.pins[output_channel].value = bool(not self.options_channels['on_state'][output_channel])
+                self.output_states[output_channel] = bool(not self.options_channels['on_state'][output_channel])
+                self.logger.debug("Output channel {} turned OFF".format(output_channel))
 
-            self.logger.debug("List sent to device: {}".format(self.dict_to_list_states(dict_states)))
-            self.sensor.port(self.dict_to_list_states(dict_states))
-            self.output_states[output_channel] = dict_states[output_channel]
             msg = "success"
         except Exception as e:
             msg = "CH{} state change error: {}".format(output_channel, e)
@@ -279,45 +209,12 @@ class OutputModule(AbstractOutput):
     def is_setup(self):
         return self.output_setup
 
-    @staticmethod
-    def dict_to_list_states(dict_states):
-        list_states = []
-        for i, _ in enumerate(dict_states):
-            list_states.append(dict_states[i])
-        return list_states
-
     def stop_output(self):
         """ Called when Output is stopped """
-        dict_states = {}
         if self.is_setup():
             for channel in channels_dict:
                 if self.options_channels['state_shutdown'][channel] == 1:
-                    dict_states[channel] = bool(self.options_channels['on_state'][channel])
+                    self.pins[channel].value = bool(self.options_channels['on_state'][channel])
                 elif self.options_channels['state_shutdown'][channel] == 0:
-                    dict_states[channel] = bool(not self.options_channels['on_state'][channel])
-            self.logger.debug("List sent to device: {}".format(self.dict_to_list_states(dict_states)))
-            self.sensor.port(self.dict_to_list_states(dict_states))
+                    self.pins[channel].value = bool(not self.options_channels['on_state'][channel])
         self.running = False
-
-
-class PCF8574(object):
-    """ A software representation of a single PCF8574 IO expander chip """
-    def __init__(self, smbus, i2c_bus, i2c_address):
-        self.bus_no = i2c_bus
-        self.bus = smbus.SMBus(i2c_bus)
-        self.address = i2c_address
-
-    def __repr__(self):
-        return "PCF8574(i2c_bus_no=%r, address=0x%02x)" % (self.bus_no, self.address)
-
-    def port(self, value):
-        """ Set the whole port using a list """
-        if not isinstance(value, list):
-            raise AssertionError
-        if len(value) != 8:
-            raise AssertionError
-        new_state = 0
-        for i, val in enumerate(value):
-            if val:
-                new_state |= 1 << i
-        self.bus.write_byte(self.address, new_state)